requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3 